Deprecate DateTimeImmutable
Fixes #1926
